Fix allocation size bug when size_t != uint64_t
(cherry picked from commit b64ab53961053f4692acf6ddac95e8dd11c77dc4)

+++ contrib/hxt/tetMesh/src/hxt_tetRefine.c
@@ -302,7 +302,7 @@ static HXTStatus balanceRefineWork(HXTMesh* mesh, uint32_t* startPt, uint64_t* s
 {
   uint64_t* scan;
   uint32_t ptPerThreadGoal;
-  HXT_CHECK( hxtAlignedMalloc(&scan, sizeof(size_t) * mesh->tetrahedra.num) );
+  HXT_CHECK( hxtAlignedMalloc(&scan, sizeof(uint64_t) * mesh->tetrahedra.num) );
 
   #pragma omp parallel num_threads(maxThreads)
   {
